Fix unset xattr on darwin
